test(flow): split compaction projection coverage
